Fixed jpackage issue using joou for Java 9+
Using jpackage lead to an error because there was a dependency to joou-java-6. By introducing two profiles this issue could be solved. The correct profile is activated based on the Java environment (defined through the JAVA_HOME environment variable). For Java 9 and later joou for Java 9 and later is used while for Java 8 and earlier joou-java-6 is used as dependency. Tested with Java 8 and Java 11